+++ index.js
@@ -189,7 +189,7 @@ class NuHeatPlatform {
                 entry = { uuid };
                 this.accessories.push(entry);
             }
-            entry.accessory = new NuHeatThermostat(this.log, deviceData, this.config.holdLength || 1440, getPlatformAccessory(deviceAccessory), this.NuHeatAPI, Homebridge);
+            entry.accessory = new NuHeatThermostat(this.log, deviceData, this.config.holdLength || 1440, getPlatformAccessory(deviceAccessory), this.NuHeatAPI, Homebridge, this.account?.temperatureScale);
             entry.existsInConfig = true;
             this.log.info("Loaded thermostat " +
                 deviceData.serialNumber +